Новый пост отправляется на сервер но не добавляется в память сразу.

+++ app/src/main/java/ru/maxpek/friendslinkup/repository/post/PostRepository.kt
@@ -11,11 +11,11 @@ interface PostRepository {
     val dataUsersMentions: MutableLiveData<List<UserRequested>>
     suspend fun loadUsersMentions(list: List<Int>)
     suspend fun getAll()
-    fun getNewerCount(id: Long): Flow<Int>
+    fun getNewerCount(id: Int): Flow<Int>
     suspend fun save(postResponse: PostResponse)
-    suspend fun removeById(id: Long)
-    suspend fun likeById(id: Long)
-    suspend fun disLikeById(id: Long)
+    suspend fun removeById(id: Int)
+    suspend fun likeById(id: Int)
+    suspend fun disLikeById(id: Int)
     suspend fun update()
-    suspend fun isSize(): Long
+    suspend fun isSize(): Int
 }
